% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{det.matrix}
\alias{det.matrix}
\title{Default implementation for the determinant of a matrix.}
\usage{
\method{det}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix}
}
\description{
Default implementation for the determinant of a matrix.
}
