% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{det.numeric}
\alias{det.numeric}
\title{Default implementation for the determinant of a number.}
\usage{
\method{det}{numeric}(x, ...)
}
\arguments{
\item{x}{a number}
}
\description{
Default implementation for the determinant of a number.
}
