% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{diag.matrix}
\alias{diag.matrix}
\title{Default implementation for the diagonal of a matrix.}
\usage{
\method{diag}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix}
}
\description{
Default implementation for the diagonal of a matrix.
}
