% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{diag.numeric}
\alias{diag.numeric}
\title{Default implementation for the diagonal of a number.}
\usage{
\method{diag}{numeric}(x, ...)
}
\arguments{
\item{x}{a number}
}
\description{
Default implementation for the diagonal of a number.
}
