% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gradient.default}
\alias{gradient.default}
\title{Default implementation for the gradient of a function.}
\usage{
\method{gradient}{default}(f, ...)
}
\arguments{
\item{f}{a function}

\item{...}{additional arguments (not used)}
}
\description{
Default implementation for the gradient of a function.
}
