% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gradient.expression}
\alias{gradient.expression}
\title{Default implementation of gradient for an expression.}
\usage{
\method{gradient}{expression}(f, ...)
}
\arguments{
\item{f}{an expression}

\item{...}{additional arguments}
}
\description{
Default implementation of gradient for an expression.
}
