% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gradient.function}
\alias{gradient.function}
\title{Default implementation of gradient for a function.}
\usage{
\method{gradient}{`function`}(f, ...)
}
\arguments{
\item{f}{a function}

\item{...}{additional arguments}
}
\description{
Default implementation of gradient for a function.
}
