% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{gradient}
\alias{gradient}
\title{Gradient of a function}
\usage{
gradient(f, ...)
}
\arguments{
\item{f}{a function}

\item{...}{additional arguments passed to \code{f}. you
may, for instance, want to pass in a set
of named parameters with which to take the
gradient.}
}
\description{
Gradient of a function
}
