% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{inv.matrix}
\alias{inv.matrix}
\title{Default implementation for the inverse of a matrix.}
\usage{
\method{inv}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix}
}
\description{
Default implementation for the inverse of a matrix.
}
