% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{inv.numeric}
\alias{inv.numeric}
\title{Default implementation for the inverse of a number.}
\usage{
\method{inv}{numeric}(x, ...)
}
\arguments{
\item{x}{a number}
}
\description{
Default implementation for the inverse of a number.
}
