% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{inv.singular_matrix}
\alias{inv.singular_matrix}
\title{A singular matrix}
\usage{
\method{inv}{singular_matrix}(x, ...)
}
\arguments{
\item{x}{a singular (known to be) matrix}
}
\description{
A singular matrix
}
