% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{myderiv.call}
\alias{myderiv.call}
\title{Default implementation of the derivative of a call expression.}
\usage{
\method{myderiv}{call}(f, var, ...)
}
\arguments{
\item{f}{a call expression}

\item{var}{the derivative is taken with respect to this variable}

\item{...}{additional arguments}
}
\description{
Default implementation of the derivative of a call expression.
}
