% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{myderiv.expression}
\alias{myderiv.expression}
\title{Default implementation of the derivative of an expression.}
\usage{
\method{myderiv}{expression}(f, var, ...)
}
\arguments{
\item{f}{an expression}

\item{var}{the derivative is taken with respect to this variable}

\item{...}{additional arguments}
}
\description{
Default implementation of the derivative of an expression.
}
