% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{myderiv}
\alias{myderiv}
\title{Generic function for the derivative of an object.}
\usage{
myderiv(f, var)
}
\arguments{
\item{f}{an expression}

\item{var}{the derivative is taken with respect to this variable}

\item{...}{additional arguments}
}
\description{
Generic function for the derivative of an object.
}
