% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rank.matrix}
\alias{rank.matrix}
\title{Default implementation for the rank of a matrix.}
\usage{
\method{rank}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix}
}
\description{
Default implementation for the rank of a matrix.
}
