% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{rank.nonsingular_matrix}
\alias{rank.nonsingular_matrix}
\title{A non-singular matrix has full rank.}
\usage{
\method{rank}{nonsingular_matrix}(x, ...)
}
\arguments{
\item{x}{a non-singular (known to be) matrix}
}
\description{
A non-singular matrix has full rank.
}
