% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{trace.matrix}
\alias{trace.matrix}
\title{Default implementation for the trace of a matrix.}
\usage{
\method{trace}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix}
}
\description{
Default implementation for the trace of a matrix.
}
