% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{trace.numeric}
\alias{trace.numeric}
\title{Default implementation for the trace of a number.}
\usage{
\method{trace}{numeric}(x, ...)
}
\arguments{
\item{x}{a number}
}
\description{
Default implementation for the trace of a number.
}
\note{
The trace of a number is the number itself.
}
\seealso{
\code{\link{diag}}
}
