% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{transpose.matrix}
\alias{transpose.matrix}
\title{Default implementation for the transpose of a matrix.}
\usage{
\method{transpose}{matrix}(x, ...)
}
\arguments{
\item{x}{a matrix}
}
\description{
Default implementation for the transpose of a matrix.
}
