% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{transpose.numeric}
\alias{transpose.numeric}
\title{Default implementation for the transpose of a number.}
\usage{
\method{transpose}{numeric}(x, ...)
}
\arguments{
\item{x}{a number}
}
\description{
Default implementation for the transpose of a number.
}
