% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{transpose}
\alias{transpose}
\title{Generic function for the transpose of an object.}
\usage{
transpose(x, ...)
}
\arguments{
\item{x}{an object}
}
\description{
Generic function for the transpose of an object.
}
