% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{walk2}
\alias{walk2}
\title{walk the expression tree, not a generic method, a test case}
\usage{
walk2(e, var)
}
\arguments{
\item{e}{an expression}

\item{var}{find this one and highlight!}
}
\description{
walk the expression tree, not a generic method, a test case
}
